Move resource Supervisor and Manager interfaces to resource package
As the resource managers now return their struct instead of the Manager
interface there is no longer a dependency cycle when keeping these
interfaces in the resource package. Keeping them there allow for clearer
usage (resource.Supervisor and resource.Manager), as originally
intended.
